requirements.txt
bitsandbytes==0.43.3
langchain_community==0.3.0
langchain_core==0.3.5
langchain_openai==0.2.0
nltk==3.9.1
pandas==2.2.3
peft==0.12.0
PyYAML==6.0.2
pyyaml-include=2.1
rank_bm25==0.2.2
scikit_learn==1.5.2
torch==2.4.1
transformers==4.44.2
